pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "encrypted-json-field"
version = "1.0.0"
authors = [
  { name="El Majidi", email="elm.majidi@gmail.com" },
]
description = "Encrypted JSON field supports automatic encryption/ decryption of JSON fields with padding to conceal start and end of value."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pycryptodome==3.16.0",
    "Django>=4.2",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
]

[project.urls]
"Homepage" = "https://github.com/MurphyAdam/encrypted-json-field"
"Bug Tracker" = "https://github.com/MurphyAdam/encrypted-json-field/issues"